requirements.txt
# Core Framework
langgraph==0.2.0
langchain==0.2.0
langchain-openai==0.1.8
langchain-core==0.2.0

# Azure Integration
azure-functions==1.18.0
azure-cosmos==4.6.0
azure-identity==1.15.0
azure-keyvault-secrets==4.8.0
azure-monitor-opentelemetry==1.3.0
opentelemetry-instrumentation-fastapi==0.45b0

# Web Framework
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.3
pydantic-settings==2.2.1

# External APIs
httpx==0.27.0
stripe==8.5.0
requests==2.31.0

# Vector Store / RAG
azure-search-documents==11.4.0
tiktoken==0.6.0

# Utilities
python-dotenv==1.0.1
pyyaml==6.0.1
tenacity==8.2.3

# Testing
pytest==8.1.1
pytest-asyncio==0.23.5
pytest-cov==4.1.0
responses==0.25.0
pytest-mock==3.12.0

# Development
black==24.2.0
flake8==7.0.0
mypy==1.9.0
